Extracts authentication into a component
